CNN-AI/requirements.txt
Flask
Flask-CORS
Pillow
numpy
scikit-learn
tensorflow
